# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/hotel/{id}": {
            "get": {
                "summary": "Get hotel details and increment visit count",
                "description": "Retrieves details of a hotel by its ID and increments the visit count. Optionally, you can include a `user_id` query parameter.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        },
                        "description": "ID of the hotel to retrieve"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "12345"
                        },
                        "description": "Optional user ID to include additional draft details"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel details and updated visit count",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "amenities": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "location": {
                                            "type": "string"
                                        },
                                        "rating": {
                                            "type": "number",
                                            "format": "decimal"
                                        },
                                        "visits_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "example": {
                                        "id": 1,
                                        "name": "Hotel Sunshine",
                                        "amenities": "Free WiFi, Breakfast",
                                        "description": "A beautiful hotel in the city center.",
                                        "location": "New York",
                                        "rating": 4.5,
                                        "visits_count": 10
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hotel not found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/hotel/checkout": {
            "post": {
                "summary": "Checkout and book a hotel",
                "description": "Creates a new booking entry and increments the booking count for a hotel.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 456
                                    }
                                },
                                "required": [
                                    "user_id",
                                    "hotel_id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Checkout successful"
                                        },
                                        "booking_id": {
                                            "type": "string",
                                            "example": "bfc82a65-5c48-4b14-8c71-57f5d0a5c674"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, missing user_id or hotel_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user_id and hotel_id are required"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found, invalid user_id or hotel_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid user_id"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hotel/{hotel_id}/draft": {
            "get": {
                "summary": "Retrieve the draft for a specific user and hotel",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Draft retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "draft": {
                                            "type": "string",
                                            "example": "This is a draft string."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No draft found for the given user_id and hotel_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No draft found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, user_id is missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user_id is required"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Save or update a draft for a specific user and hotel",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "draft": {
                                        "type": "string",
                                        "example": "This is a draft string."
                                    }
                                },
                                "required": [
                                    "draft"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Draft saved or updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Draft saved successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, user_id is missing or draft is empty",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user_id is required"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/hotel/bookings": {
            "get": {
                "summary": "Retrieve bookings information for a specific hotel",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookings information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "total_visits": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "completed_bookings": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total_draft_bookings": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, hotel_id is missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "hotel_id is required"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid hotel_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid hotel_id"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/suggest": {
            "get": {
                "summary": "Suggest hotels based on user visit history or top ratings",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of suggested hotels",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Hotel A"
                                            },
                                            "amenities": {
                                                "type": "string",
                                                "example": "WiFi, Pool, Gym"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "A luxurious hotel in city center"
                                            },
                                            "location": {
                                                "type": "string",
                                                "example": "New York"
                                            },
                                            "rating": {
                                                "type": "string",
                                                "example": "4.50"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}